+++ ITMLib/Engines/Visualisation/CUDA/ITMVisualisationHelpers_CUDA.cu
@@ -130,7 +130,7 @@ __global__ void ITMLib::findMissingPoints_device(int *fwdProjMissingPoints, uint
 	if (x >= imgSize.x || y >= imgSize.y) return;
 
 	int locId = x + y * imgSize.x;
-	int locId2 = (int)floor((float)x / minmaximg_subsample) + (int)floor((float)y / minmaximg_subsample) * imgSize.x;
+	int locId2 = (int)floorf((float)x / minmaximg_subsample) + (int)floorf((float)y / minmaximg_subsample) * imgSize.x;
 
 	Vector4f fwdPoint = forwardProjection[locId];
 	Vector2f minmaxval = minmaximg[locId2];
